relocate: RemindFinePushNotificationGenerator
